requirements.txt
confluent_kafka>=2.4,<2.5
Django==3.1.12
django-cors-headers==3.5.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djongo==1.3.7
dnspython==2.7.0
elasticsearch==8.17.1
firebase-admin==6.6.0
huggingface_hub==0.28.1
langchain==0.3.17
loguru==0.7.3
pymongo==3.11.4
python-dotenv==1.0.1
quixstreams==3.7.0
redis==5.2.1
rest_framework_simplejwt==0.0.2
setuptools==75.8.0
